// src/app/store/user/types.ts
export enum userActionTypes {
  CHANGE_ONLINE_STATUS = 'CHANGE_ONLINE_STATUS',
  CHANGE_INTERLOCUTOR_STATUS = 'CHANGE_INTERLOCUTOR_STATUS',
  USER_STATUS_CHANGED_EVENT = 'USER_STATUS_CHANGED_EVENT',
  UPDATE_MY_PROFILE_INFO = 'UPDATE_MY_PROFILE_INFO',
  UPDATE_MY_AVATAR = 'UPDATE_MY_AVATAR',
  GET_MY_PROFILE = 'GET_MY_PROFILE',
  GET_MY_PROFILE_SUCCESS = 'GET_MY_PROFILE_SUCCESS',
  UPDATE_MY_AVATAR_SUCCESS = 'UPDATE_MY_AVATAR_SUCCESS'
}