// src/app/store/conferences/types.ts
export enum ConferencesActionTypes {
  CREATE_CONFERENCE = 'CREATE_CONFERENCE',
  CREATE_CONFERENCE_SUCCESS = 'CREATE_CONFERENCE_SUCCESS',
  CREATE_CONFERENCE_FROM_EVENT = 'CREATE_CONFERENCE_FROM_EVENT',
  CONFERENCE_MESSAGE_READ_FROM_EVENT = 'CONFERENCE_MESSAGE_READ_FROM_EVENT',

  GET_CONFERENCE_USERS = 'GET_CONFERENCE_USERS',
  GET_CONFERENCE_USERS_SUCCESS = 'GET_CONFERENCE_USERS_SUCCESS',
  UNSET_CONFERENCE_USERS = 'UNSET_CONFERENCE_USERS',

  UNSET_SELECTED_USER_IDS_TO_ADD_INTO_CONFERENCE = 'UNSET_SELECTED_USER_IDS_TO_ADD_INTO_CONFERENCE',

  LEAVE_CONFERENCE = 'LEAVE_CONFERENCE',
  LEAVE_CONFERENCE_SUCCESS = 'LEAVE_CONFERENCE_SUCCESS',
  ADD_USERS_TO_CONFERENCE = 'ADD_USERS_TO_CONFERENCE',
  ADD_USERS_TO_CONFERENCE_SUCCESS = 'ADD_USERS_TO_CONFERENCE_SUCCESS',

  RENAME_CONFERENCE = 'RENAME_CONFERENCE',
  RENAME_CONFERENCE_SUCCESS = 'RENAME_CONFERENCE_SUCCESS',

  CHANGE_CONFERENCE_AVATAR = 'CHANGE_CONFERENCE_AVATAR',
  CHANGE_CONFERENCE_AVATAR_SUCCESS = 'CHANGE_CONFERENCE_AVATAR_SUCCESS'
}